feat: add db push command that skips the migration trail
